Update Derpibooru support, Add other Boorus
This is a collection of fixes and updates from anons in the 8chan thread for the parser, along with my own fixes to update support for Derpibooru's new tag changes. This version of the code also downloads the short ID named version of Derpibooru images rather than the long named version. Support is also added for SoFurry tags, though the parser might have trouble with some of them.